backend/requirements.txt
fastapi
uvicorn
python-dotenv

google-genai
psycopg2-binary
